Rework `get_flat_value` to fix node relationships (#487)
This change moves the `get_flat_value` function to methods of the `InfrahubNode` and `InfrahubNodeSync` classes. This allows to handle traversing relationships of cardinality one using a flat notation in addition to attributes.

The `extract` method has also been moved as it needs to be async for regular nodes now.