Add `value_to_enum_name` filter
This Jinja2 filter aims to convert a value into a name when it is part
of an enum. This is useful when the name of an enum member has more
meaning, e.g. in a string which is exposed to users.

An optional `enum_path` parameter can be specified. When using a
valid import string for an enum class, that class will be imported to
look for the value within the enum members. This mean that this filter
can be used with any enum classes, including the ones found in the
Infrahub server code.